requirements.txt
Flask
Flask-SQLAlchemy
Flask-Security-Too
SQLAlchemy-serializer
mysqlclient
bcrypt
rfc3339
ansicolors
bibtexparser
pillow
gunicorn